+++ sr_ronex_transmissions/src/mapping/general_io/command_to_pwm.cpp
@@ -48,6 +48,21 @@ namespace ronex
 
       bool CommandToPWM::try_init_cb_(const ros::TimerEvent&, TiXmlElement* mapping_el, ros_ethercat_model::RobotState* robot, const char* ronex_name)
       {
+        if( !init_(mapping_el, robot, ronex_name) )
+        {
+          return false;
+        }
+
+        ROS_DEBUG_STREAM("RoNeX" << ronex_name << " is initialised now.");
+        //stopping timer
+        init_timer_.stop();
+
+        is_initialized_ = true;
+        return true;
+      }
+
+      bool CommandToPWM::init_(TiXmlElement* mapping_el, ros_ethercat_model::RobotState* robot, const char* ronex_name)
+      {
         //has the ronex been added by the driver?
         if( robot->getCustomHW(ronex_name) == NULL )
           return false;
@@ -113,16 +128,7 @@ namespace ronex
           return false;
         }
 
-        ROS_DEBUG_STREAM("RoNeX" << ronex_name << " is initialised now.");
-        //stopping timer
-        init_timer_.stop();
-
-        is_initialized_ = true;
         return true;
-      }
-
-      CommandToPWM::~CommandToPWM()
-      {
       }
 
       bool CommandToPWM::check_pins_in_bound_()
@@ -141,7 +147,7 @@ namespace ronex
           if( pwm_module_ >= general_io_->command_.pwm_.size() )
           {
             //size_t is always >= 0 so no need to check lower bound
-            ROS_ERROR_STREAM("Specified PWM module index is out of bound: " << pwm_pin_index_ << " / max = " << general_io_->command_.pwm_.size() << ", not propagating the command to the RoNeX.");
+            ROS_ERROR_STREAM("Specified PWM module index is out of bound: " << pwm_module_ << " / max = " << general_io_->command_.pwm_.size() << ", not propagating the command to the RoNeX.");
             pin_out_of_bound_ = true;
             return false;
           }
@@ -165,22 +171,20 @@ namespace ronex
         return true;
       }
 
-      void CommandToPWM::propagateToRonex(std::vector<ros_ethercat_model::JointState*>& js)
+      void CommandToPWM::propagateToRonex(ros_ethercat_model::JointState *js)
       {
         if( !is_initialized_ )
           return;
 
-        assert(js.size() == 1);
-
         if( check_pins_in_bound_() )
         {
           if( pwm_pin_index_ == 0 )
-            general_io_->command_.pwm_[pwm_module_].on_time_0 = static_cast<unsigned short int>((static_cast<double>(general_io_->command_.pwm_[pwm_module_].period) * js[0]->commanded_effort_ ) / 100);
+            general_io_->command_.pwm_[pwm_module_].on_time_0 = static_cast<unsigned short int>((static_cast<double>(general_io_->command_.pwm_[pwm_module_].period) * abs(js->commanded_effort_) ) / 100);
           else
-            general_io_->command_.pwm_[pwm_module_].on_time_1 = static_cast<unsigned short int>((static_cast<double>(general_io_->command_.pwm_[pwm_module_].period) * js[0]->commanded_effort_ ) / 100);
+            general_io_->command_.pwm_[pwm_module_].on_time_1 = static_cast<unsigned short int>((static_cast<double>(general_io_->command_.pwm_[pwm_module_].period) * abs(js->commanded_effort_) ) / 100);
 
           // This is assigned by convention: negative effort sets the direction pin to 1.
-          general_io_->command_.digital_[digital_pin_index_] = (js[0]->commanded_effort_ < 0.0);
+          general_io_->command_.digital_[digital_pin_index_] = (js->commanded_effort_ < 0.0);
        }
       }
     }