+++ sr_ronex_transmissions/src/ronex_transmission.cpp
@@ -1,16 +1,16 @@
 /*
  * Copyright (c) 2013, Shadow Robot Company, All rights reserved.
- * 
+ *
  * This library is free software; you can redistribute it and/or
  * modify it under the terms of the GNU Lesser General Public
  * License as published by the Free Software Foundation; either
  * version 3.0 of the License, or (at your option) any later version.
- * 
+ *
  * This library is distributed in the hope that it will be useful,
  * but WITHOUT ANY WARRANTY; without even the implied warranty of
  * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU
  * Lesser General Public License for more details.
- * 
+ *
  * You should have received a copy of the GNU Lesser General Public
  * License along with this library.
  */
@@ -30,6 +30,7 @@
 #include "sr_ronex_transmissions/mapping/general_io/analogue_to_position.hpp"
 #include "sr_ronex_transmissions/mapping/general_io/analogue_to_effort.hpp"
 #include "sr_ronex_transmissions/mapping/general_io/command_to_pwm.hpp"
+#include "sr_ronex_transmissions/mapping/general_io/command_to_pwm_2_dir_pin.hpp"
 
 PLUGINLIB_EXPORT_CLASS( ronex::RonexTransmission, ros_ethercat_model::Transmission)
 
@@ -39,6 +40,21 @@ namespace ronex
   {
     if (!ros_ethercat_model::Transmission::initXml(elt, robot))
       return false;
+
+    //reading the joint name
+    TiXmlElement *jel = elt->FirstChildElement("joint");
+    if (!jel || !jel->Attribute("name"))
+    {
+      ROS_ERROR_STREAM("Joint name not specified in transmission " << name_);
+      return false;
+    }
+
+    joint_ = robot->getJointState(jel->Attribute("name"));
+
+    // TODO Modify the ros_ethercat_model so that a transmission is not expected to have an actuator
+    // (currently the ronex mappings don't use actuators, they access a ronex::GeneralIO object directly)
+    actuator_ = new ros_ethercat_model::Actuator();
+    actuator_->name_ = "Dummy_actuator_" + std::string(jel->Attribute("name"));
 
     //Extract all the mapping information from the transmission
     for( TiXmlElement *mapping_el = elt->FirstChildElement("mapping"); mapping_el;
@@ -64,29 +80,31 @@ namespace ronex
         {
           ronex_mappings_.push_back( new mapping::general_io::CommandToPWM(mapping_el, robot) );
         }
+        else if( std::strcmp("command_2_dir", property) == 0 )
+        {
+          ronex_mappings_.push_back( new mapping::general_io::CommandToPWM2PinDir(mapping_el, robot) );
+        }
         else
-          ROS_WARN_STREAM("Property not recognised: " << property);
+          ROS_WARN_STREAM("Property not recognized: " << property);
       }
     }
 
     return true;
   }
 
-  void RonexTransmission::propagatePosition(std::vector<ros_ethercat_model::Actuator*>& as,
-                                            std::vector<ros_ethercat_model::JointState*>& js)
+  void RonexTransmission::propagatePosition()
   {
     for(ronex_iter_ = ronex_mappings_.begin(); ronex_iter_ != ronex_mappings_.end(); ++ronex_iter_)
     {
-      ronex_iter_->propagateFromRonex(js);
+      ronex_iter_->propagateFromRonex(joint_);
     }
   }
 
-  void RonexTransmission::propagateEffort(std::vector<ros_ethercat_model::JointState*>& js,
-                                          std::vector<ros_ethercat_model::Actuator*>& as)
+  void RonexTransmission::propagateEffort()
   {
     for(ronex_iter_ = ronex_mappings_.begin(); ronex_iter_ != ronex_mappings_.end(); ++ronex_iter_)
     {
-      ronex_iter_->propagateToRonex(js);
+      ronex_iter_->propagateToRonex(joint_);
     }
   }
 }